Move CDiskBlockPos from chain to flatfile.
